requirements.txt
fastapi
uvicorn
pydantic
jinja2
python-multipart
openai
firebase-admin
PyPDF2
python-docx
Pillow
requests
python-dotenv
httpx
starlette
